CUC_Report_by_Requests/requirements.txt
requests
PyExecJS
bs4
